Rename JsonViews and DemoMode
