+++ src/main/java/com/example/jsonview/JsonViews.java
@@ -1,6 +1,6 @@
-package com.example.view;
+package com.example.jsonview;
 
-public class Views {
+public class JsonViews {
     public interface ProductBasic {} //ProductBasic
     public interface ImageBasic {} //ImageBasic
 }
